Index Addition (#3)
* Created index.js

Not the final version. Several possible changes pending.

* Update index.js

Fixed a small error.

* Simplification

Removed the error objects, as I realized in hindsight that this probably isn't the right file for them. Also removed some redundant array properties from a couple of the objects. Customized the property fields the professor provided as a starter to be something a little more amusing.

* Update index.js